requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
frozenlist==1.3.3
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
idna==3.4
kombu==5.2.4
Light-Pipe==0.2.2
multidict==6.0.4
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.4
requests==2.29.0
rsa==4.9
six==1.16.0
torch==1.11.0
torchvision==0.12.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
